Resume templates restructure
