+++ src/components/pages/Resume.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
 import { Flex } from "../common/styles";
 import { useTemplates } from "../stores/settings.store";
