+++ src/resume-templates/items/intro/Intro.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import { Flex, FlexCol, FlexVC } from "../../../common/styles";
-import { getIcon } from "../../../common/icons";
+import { Flex, FlexCol, FlexVC } from "../../../components/common/styles";
+import { getIcon } from "../../../components/common/icons";
 import Color from "color";
 
 const Role = styled.h3`
