+++ src/resume-templates/items/separator/LineSeparator.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const Separator = styled.div`
+export const LineSeparator = styled.div`
   width: 80%;
   margin: 10px auto 4px auto;
   min-height: 1px;
