+++ src/resume-templates/items/skills/RatedBars.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { FlexCol, FlexVC } from "../../../common/styles";
+import { FlexCol, FlexVC } from "../../../components/common/styles";
 import { Progress } from "antd";
 import { withTheme } from "styled-components";
 
@@ -25,7 +25,7 @@ function RatedElementComponent({ itemLabel, itemRating, theme }: any) {
 
 const RatedElement = withTheme(RatedElementComponent);
 
-export function RatedSection({ items }: any) {
+export function RatedBars({ items }: any) {
   return (
     <FlexCol rGap="8px">
       {items.map((data: any, index: number) => (
